Added reference timestamp and schedule for "now"

+++ src/main/java/org/radarbase/appserver/client/protocol/RepeatQuestionnaire.kt
@@ -1,6 +1,6 @@
 package org.radarbase.appserver.client.protocol
 
 data class RepeatQuestionnaire(
-    val unitsFromZero: List<Int> = listOf(),
+    val unitsFromZero: List<Long> = listOf(),
     val unit: String = "min",
 )
